requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
frozenlist==1.3.3
google==3.0.0
googlesearch-python==1.2.3
idna==3.4
ipykernel==6.22.0
ipython==8.13.1
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
lxml==4.9.2
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
openai==0.27.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
python-docx==0.8.11
PyYAML==6.0
pyzmq==25.0.2
requests==2.29.0
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.9.2
